Fix compiler warnings

+++ hit.h
@@ -595,7 +595,7 @@ protected:
 	size_t nthreads_;
 	std::vector<BTString> ptBufs_;
 	std::vector<size_t> ptCounts_;
-	int perThreadBufSize_;
+	size_t perThreadBufSize_;
 	bool reorder_;
 
 	// Output filenames for dumping
